feat: merged Aspen's commits into upload component
