+++ ui/src/components/UploadStory.tsx
@@ -1,6 +1,6 @@
 import {
+  AppBar,
   Button,
-  Checkbox,
   Chip,
   Dialog,
   DialogActions,
@@ -19,18 +19,17 @@ import {
 } from "@material-ui/core/";
 import Alert from "@material-ui/lab/Alert";
 import Autocomplete from "@material-ui/lab/Autocomplete";
-import { countReset } from "console";
 import { DropzoneArea } from "material-ui-dropzone";
 import * as React from "react";
 import { KeyboardEvent, useReducer, useState } from "react";
 import styled from "styled-components";
 import useSWR from "swr";
 
-import countriesList from "../data/countries.json";
+import { StoryDrawer } from "../components";
 import { colors } from "../styles/colors";
 import { device } from "../styles/device";
 import { UploadLabelsText, UploadStoriesHeading } from "../styles/typography";
-import { Story } from "../types";
+import { Author, Story } from "../types";
 
 const StyledGrid = styled(Grid)`
   background-color: ${colors.primaryLight6};
@@ -85,6 +84,20 @@ const StyledDropzoneArea = styled(DropzoneArea)`
   }
 `;
 
+const StyledTags = styled(Autocomplete)`
+  && {
+    font-family: Poppins;
+    margin-top: 12px;
+    width: 30vw;
+    .MuiInputLabel-root {
+      font-family: Poppins !important;
+    }
+    .MuiInputBase-input {
+      font-family: Poppins !important;
+    }
+  }
+`;
+
 const StyledAutocomplete = styled(Autocomplete)`
   && {
     font-family: Poppins;
@@ -100,10 +113,8 @@ const StyledAutocomplete = styled(Autocomplete)`
 `;
 
 const StyledButton = styled(Button)`
-  && {
-    width: 30vw;
-    margin-left: 24px;
-  }
+  text-transform: none;
+  margin: 5px;
 `;
 
 const AddCountryPaper = styled(Paper)`
@@ -154,28 +165,29 @@ interface TagParameters {
 
 interface StoryProps {
   id: number;
-  story: Story;
+  currentStory: Story;
   bio: string;
 }
 
 export const UploadStory: React.FC<StoryProps> = ({
   id,
-  story,
+  currentStory,
   bio,
 }: StoryProps) => {
   const { data: tagOptions, error } = useSWR<string[]>("/api/tags");
-  const [tagArray, setTagArrayValues] = useState(story.tags);
-  const [authorCountry, setAuthorCountry] = useState(story.author_country);
-  const [autocompleteAuthor, setAutocompleteAuthor] = useState(
-    story.author_country
+  const [tagArray, setTagArrayValues] = useState(currentStory.tags);
+  const [authorCountry, setAuthorCountry] = useState(currentStory.author_country);
+  const [autocompleteAuthorCountry, setAutocompleteAuthorCountry] = useState(
+    currentStory.author_country
   );
   const [addedCountry, setAddedCountry] = useState("");
-  const [newImage, setNewImage] = useState(story.image_url);
+  const [newImage, setNewImage] = useState(currentStory.image_url);
   const [disabled, setDisabled] = useState(false);
   //handleSubmit component states
   const [dialogOpenState, setDialogOpenState] = useState(false);
   const [uploadErrorState, setErrorOpen] = useState(false);
   const [loading, setLoading] = useState(false);
+  const [isDrawerOpen, setIsDrawerOpen] = useState(false);
 
   const { data: countries, error: errorCountries } = useSWR<string[]>(
     "/api/countries"
@@ -196,18 +208,75 @@ export const UploadStory: React.FC<StoryProps> = ({
     (state, newState) => ({ ...state, ...newState }),
     {
       image: new File([""], "filename"),
-      video_url: story.video_url,
-      title: story.title,
-      content: story.content,
-      summary: story.summary,
-      author_first_name: story.author_first_name,
-      author_last_name: story.author_last_name,
-      // author_country: story.author_country,
-      year: story.year,
-      current_city: story.current_city,
+      video_url: currentStory.video_url,
+      title: currentStory.title,
+      content: currentStory.content,
+      summary: currentStory.summary,
+      author_first_name: currentStory.author_first_name,
+      author_last_name: currentStory.author_last_name,
+      author_country: currentStory.author_country,
+      year: currentStory.year,
+      current_city: currentStory.current_city,
       bio: bio,
     }
   );
+
+  const story = React.useMemo(() => {
+    if (isDrawerOpen) {
+      const storyFromData: Story = {
+        image_url: (newImage == "" ? URL.createObjectURL(formInput.image) : currentStory.image_url),
+        video_url: formInput.video_url as string,
+        title: formInput.title as string,
+        content: formInput.content as string,
+        summary: formInput.summary as string,
+        author_first_name: formInput.author_first_name as string,
+        author_last_name: formInput.author_last_name as string,
+        author_country: authorCountry as string,
+        year: formInput.year as number,
+        current_city: formInput.current_city as string,
+        author: {
+          ID: 0,
+          CreatedAt: "",
+          UpdatedAt: "",
+          DeletedAt: {
+            Time: "",
+            Valid: true,
+          },
+          firstName: formInput.author_last_name as string,
+          currentCity: formInput.current_city as string,
+          bio: formInput.bio as string,
+        } as Author,
+        tags: [],
+        is_visible: true,
+        latitude: 0,
+        longitude: 0,
+        ID: 0,
+        CreatedAt: "",
+        UpdatedAt: "",
+        DeletedAt: {
+          Time: "",
+          Valid: true,
+        },
+      };
+      return storyFromData;
+    }
+    return undefined;
+  }, [isDrawerOpen, formInput]);
+
+  const hasAllRequiredFields = React.useMemo(() => {
+
+    return (
+      formInput.image &&
+      formInput.title &&
+      formInput.content &&
+      formInput.summary &&
+      formInput.author_first_name &&
+      formInput.author_last_name &&
+      formInput.year &&
+      formInput.current_city &&
+      formInput.author_country
+    );
+  }, [formInput]);
 
   const setImage = (files: File[]) => {
     setFormInput({
@@ -272,26 +341,29 @@ export const UploadStory: React.FC<StoryProps> = ({
     }
   };
 
-  const addCountryButtonPressed = (autocompleteAuthor: string) => {
-    setAddedCountry(autocompleteAuthor);
-    setAuthorCountry(autocompleteAuthor);
-    countries.push(autocompleteAuthor);
+  const addCountryButtonPressed = (autocompleteAuthorCountry: string) => {
+    setFormInput({
+      ["author_country"]: autocompleteAuthorCountry,
+    });
+    setAddedCountry(autocompleteAuthorCountry);
+    setAuthorCountry(autocompleteAuthorCountry);
+    countries.push(autocompleteAuthorCountry);
   };
 
-  const newCountry = ({ children, ...other }) => (
+  const addNewCountry = ({ children, ...other }) => (
     <AddCountryPaper {...other}>
       {countries.filter(
-        (str) => str.toLowerCase() == autocompleteAuthor.toLowerCase()
+        (str) => str.toLowerCase() == autocompleteAuthorCountry.toLowerCase()
       ).length == 0 &&
-        autocompleteAuthor != "" && (
+      autocompleteAuthorCountry != "" && (
           <AddCountryDiv
             onMouseDown={(event) => {
               event.preventDefault();
             }}
           >
-            {autocompleteAuthor}
+            {autocompleteAuthorCountry}
             <AddCountryButton
-              onClick={() => addCountryButtonPressed(autocompleteAuthor)}
+              onClick={() => addCountryButtonPressed(autocompleteAuthorCountry)}
             >
               ADD
             </AddCountryButton>
@@ -338,7 +410,7 @@ export const UploadStory: React.FC<StoryProps> = ({
     let method = "POST";
 
     if (id) {
-      // if a new image was uplaoded, add the image to the formdata
+      // if a new image was uploaded, add the image to the formdata
       if (newImage == "") {
         formData.append("image", formInput["image"]);
       }
@@ -350,13 +422,13 @@ export const UploadStory: React.FC<StoryProps> = ({
 
     formData.append("author_country", authorCountry);
     if (authorCountry === addedCountry) {
-      let empArray = [
+      let countryNameJsonBody = [
         {
           country_name: addedCountry,
         },
       ];
 
-      const jsonBody = JSON.stringify(empArray);
+      const jsonBody = JSON.stringify(countryNameJsonBody);
       fetch("/api/countries", {
         method: "POST",
         body: jsonBody,
@@ -370,12 +442,46 @@ export const UploadStory: React.FC<StoryProps> = ({
   };
 
   const successMessage = id ? "Story Edit Success!" : "Story Upload Success!";
+  const pageTitle = id ? "Edit Story" : "Upload New Story";
+  const submitButtonText = id ? "Save Edits" : "Upload";
 
   if (error) return <div>Error fetching tags!</div>;
   if (errorCountries) return <div>Error fetching countries array!</div>;
 
   return (
     <>
+      <AppBar color="default" position="sticky">
+        <Grid container direction="row">
+          <Grid item xs={6}>
+            <UploadStoriesHeading>{pageTitle}</UploadStoriesHeading>
+          </Grid>
+          <Grid
+            container
+            item
+            xs={6}
+            direction="row"
+            alignContent="center"
+            justify="flex-end"
+          >
+            <StyledButton
+              disabled={!hasAllRequiredFields}
+              onClick={() => setIsDrawerOpen(true)}
+              color="primary"
+              variant="outlined"
+            >
+              Preview
+            </StyledButton>
+            <StyledButton
+              disabled={disabled || !hasAllRequiredFields}
+              variant="contained"
+              color="primary"
+              onClick={handleSubmit}
+            >
+              {submitButtonText}
+            </StyledButton>
+          </Grid>
+        </Grid>
+      </AppBar>
       <StyledGrid container justify="center" alignContent="center">
         <form onSubmit={handleSubmit}>
           <FormControl>
@@ -476,9 +582,13 @@ export const UploadStory: React.FC<StoryProps> = ({
                   // prettier-ignore
                   value={addedCountry != "" ? addedCountry : (authorCountry ? authorCountry : null)}
                   onInputChange={(_, newValue) =>
-                    setAutocompleteAuthor(newValue)
+                    setAutocompleteAuthorCountry(newValue)
                   }
-                  onChange={(_, newValue) => setAuthorCountry(newValue)}
+                  onChange={(_, newValue) => {
+                    setFormInput({
+                      ["author_country"]: newValue,
+                    });
+                   setAuthorCountry(newValue)}}
                   renderInput={(params) => {
                     return (
                       <TextField
@@ -489,7 +599,7 @@ export const UploadStory: React.FC<StoryProps> = ({
                       />
                     );
                   }}
-                  PaperComponent={newCountry}
+                  PaperComponent={addNewCountry}
                 />
               </FormControl>
               <UploadLabelsText>Current City</UploadLabelsText>
@@ -587,7 +697,7 @@ export const UploadStory: React.FC<StoryProps> = ({
                   <div>
                     <UploadLabelsText>Current Image:</UploadLabelsText>
                     <br />
-                    <StyledImage src={story.image_url} />
+                    <StyledImage src={currentStory.image_url} />
                     <Button
                       color="primary"
                       variant="contained"
@@ -612,14 +722,14 @@ export const UploadStory: React.FC<StoryProps> = ({
                 defaultValue={formInput.video_url}
               />
             </StyledBackgroundColor>
-            <StyledButton
+            {/* <StyledButton
               disabled={disabled}
               color="primary"
               type="submit"
               variant="contained"
             >
               Submit Story
-            </StyledButton>
+            </StyledButton> */}
             <Snackbar
               open={uploadErrorState}
               autoHideDuration={5000}
@@ -655,6 +765,7 @@ export const UploadStory: React.FC<StoryProps> = ({
           </FormControl>
         </form>
       </StyledGrid>
+      <StoryDrawer story={story} onClose={() => setIsDrawerOpen(false)} />
       {loading ? <StyledLinearProgress /> : null}
     </>
   );
